Refactor of RasterTile RasterPolygonTile into RasterTileMetadata and RasterPolygonTileMetadata, which don't contain duplicated pixel data from the raster, saving huge amount of memory. Also reformatted a bit the relevant files
