--- geodataset/geodata/base_geodata.py
@@ -1,8 +1,0 @@
-class BaseGeoData:
-    """
-    Abstract class containing base methods and attributes for any type of Geo Data (raster or point cloud)
-    """
-
-    @staticmethod
-    def _load_data():
-        pass
